API/desafio-api-clientes/api_clientes/api/urls.py
from django.urls import path, include
from django.contrib.auth.models import User
from rest_framework import routers, serializers, viewsets
from .views import Tbl_ClientesViewSet, Tbl_EnderecosViewSet


# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'clientes_api', Tbl_ClientesViewSet)
router.register(r'enderecos_api', Tbl_EnderecosViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
]