API/desafio-api-clientes/api_clientes/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
Markdown==3.3.4
mysql-connector-python==8.0.23
mysqlclient==2.0.3
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
validate-docbr==1.8.2
virtualenv==20.4.3
